requirements.txt
chess==1.11.2

fastapi==0.115.14
uvicorn==0.34.2
pydantic==2.11.4
redis==6.2.0
httpx==0.28.1

pandas==2.3.0

python-dotenv==1.1.1
google-genai==1.24.0
